requirements_gui.txt
PyQt5==5.15.10
yt-dlp==2024.1.14
ffmpeg-python==0.2.0
pyinstaller==6.3.0
pillow==10.2.0